move state methods to bevy_states
