+++ crates/bevy_app/src/sub_app.rs
@@ -1,14 +1,9 @@
-use crate::{App, InternedAppLabel, Plugin, Plugins, PluginsState, Startup};
+use crate::{App, InternedAppLabel, Plugin, Plugins, PluginsState};
 use bevy_ecs::{
     event::EventRegistry,
     prelude::*,
     schedule::{InternedScheduleLabel, ScheduleBuildSettings, ScheduleLabel},
     system::SystemId,
-};
-#[cfg(feature = "bevy_state")]
-use bevy_state::{
-    prelude::*,
-    state::{setup_state_transitions_in_world, FreelyMutableState},
 };
 
 #[cfg(feature = "trace")]
@@ -294,88 +289,6 @@ impl SubApp {
         let mut schedules = self.world.resource_mut::<Schedules>();
 
         schedules.ignore_ambiguity(schedule, a, b);
-
-        self
-    }
-
-    #[cfg(feature = "bevy_state")]
-    /// See [`App::init_state`].
-    pub fn init_state<S: FreelyMutableState + FromWorld>(&mut self) -> &mut Self {
-        if !self.world.contains_resource::<State<S>>() {
-            setup_state_transitions_in_world(&mut self.world, Some(Startup.intern()));
-            self.init_resource::<State<S>>()
-                .init_resource::<NextState<S>>()
-                .add_event::<StateTransitionEvent<S>>();
-            let schedule = self.get_schedule_mut(StateTransition).unwrap();
-            S::register_state(schedule);
-            let state = self.world.resource::<State<S>>().get().clone();
-            self.world.send_event(StateTransitionEvent {
-                exited: None,
-                entered: Some(state),
-            });
-        }
-
-        self
-    }
-
-    #[cfg(feature = "bevy_state")]
-    /// See [`App::insert_state`].
-    pub fn insert_state<S: FreelyMutableState>(&mut self, state: S) -> &mut Self {
-        if !self.world.contains_resource::<State<S>>() {
-            setup_state_transitions_in_world(&mut self.world, Some(Startup.intern()));
-            self.insert_resource::<State<S>>(State::new(state.clone()))
-                .init_resource::<NextState<S>>()
-                .add_event::<StateTransitionEvent<S>>();
-            let schedule = self.get_schedule_mut(StateTransition).unwrap();
-            S::register_state(schedule);
-            self.world.send_event(StateTransitionEvent {
-                exited: None,
-                entered: Some(state),
-            });
-        }
-
-        self
-    }
-
-    #[cfg(feature = "bevy_state")]
-    /// See [`App::add_computed_state`].
-    pub fn add_computed_state<S: ComputedStates>(&mut self) -> &mut Self {
-        if !self
-            .world
-            .contains_resource::<Events<StateTransitionEvent<S>>>()
-        {
-            setup_state_transitions_in_world(&mut self.world, Some(Startup.intern()));
-            self.add_event::<StateTransitionEvent<S>>();
-            let schedule = self.get_schedule_mut(StateTransition).unwrap();
-            S::register_computed_state_systems(schedule);
-            let state = self.world.resource::<State<S>>().get().clone();
-            self.world.send_event(StateTransitionEvent {
-                exited: None,
-                entered: Some(state),
-            });
-        }
-
-        self
-    }
-
-    #[cfg(feature = "bevy_state")]
-    /// See [`App::add_sub_state`].
-    pub fn add_sub_state<S: SubStates>(&mut self) -> &mut Self {
-        if !self
-            .world
-            .contains_resource::<Events<StateTransitionEvent<S>>>()
-        {
-            setup_state_transitions_in_world(&mut self.world, Some(Startup.intern()));
-            self.init_resource::<NextState<S>>();
-            self.add_event::<StateTransitionEvent<S>>();
-            let schedule = self.get_schedule_mut(StateTransition).unwrap();
-            S::register_sub_state_systems(schedule);
-            let state = self.world.resource::<State<S>>().get().clone();
-            self.world.send_event(StateTransitionEvent {
-                exited: None,
-                entered: Some(state),
-            });
-        }
 
         self
     }
