+++ crates/bevy_internal/Cargo.toml
@@ -183,7 +183,7 @@ bevy_dev_tools = ["dep:bevy_dev_tools"]
 ios_simulator = ["bevy_pbr?/ios_simulator", "bevy_render?/ios_simulator"]
 
 # Enable built in global state machines
-bevy_state = ["dep:bevy_state", "bevy_app/bevy_state"]
+bevy_state = ["dep:bevy_state"]
 
 [dependencies]
 # bevy
