+++ crates/bevy_state/Cargo.toml
@@ -12,14 +12,16 @@ categories = ["game-engines", "data-structures"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-default = ["bevy_reflect"]
-
+default = ["bevy_reflect", "bevy_app"]
+bevy_reflect = ["dep:bevy_reflect", "bevy_ecs/bevy_reflect"]
+bevy_app = ["dep:bevy_app"]
 
 [dependencies]
 bevy_ecs = { path = "../bevy_ecs", version = "0.14.0-dev" }
 bevy_state_macros = { path = "macros", version = "0.14.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.14.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.14.0-dev", optional = true }
+bevy_app = { path = "../bevy_app", version = "0.14.0-dev", optional = true }
 
 [lints]
 workspace = true
